requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
openpyxl
python-multipart